changed redux/actions redux/reducer some interfaces
