+++ wbooks/tsconfig.json
@@ -63,7 +63,7 @@
       "@screens/*":["src/app/screens/*"],
       "@config/*":["src/config/*"],
       "@hooks/*":["src/app/hooks/*"],
-      "@interfaces":["src/interfaces"],
+      "@interfaces/*":["src/interfaces/*"],
       "@services":["src/services/*"],
       "@utils":["src/utils"],
       "@navigator":["src/app/navigator"],
